requirements.txt
Flask>=3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai>=1.50.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.9